Separate delayed attacks from AnimatedCreatureEntity
